+++ python/TestHarness/pytest.ini
@@ -4,6 +4,9 @@ addopts =
     --cov=TestHarness.resultsstore
     --cov-config=.coveragerc
     --durations=20
-testpaths = tests
+testpaths =
+    tests/resultsstore/test_civetstore.py
+    tests/resultsstore/test_testdatafilters.py
+    tests/resultsstore/test_utils.py
 markers =
     moose: marks tests as requiring moose (use --no-moose to skip)
