--- python/TestHarness/resultsstore/tests/__init__.py
@@ -1,1 +1,0 @@
-../../../moosepytest/init_moose_python.py
