+++ python/TestHarness/tests/__init__.py
@@ -1,0 +1,1 @@
+../../moosepytest/init_moose_python.py
